[fix][#16] error, bug 수정 및 코드스타일 수정
